+++ app/__init__.py
@@ -1,2 +1,2 @@
-VERSION = "0.35.1"
+VERSION = "0.36.0"
 SYSTEM_NAME = "d4k Study Definitions Workbench"